Fixed removeEventListener behavior in respect to the options parameter
`options` no longer needs to be a reference to the same object, but simply must contain the same options.
The listener still must be a reference to the same listener, however. E.g., if a defined function passed was redefined before being passed to the removal function, it will not match and thus will not be removed.
The function also no longer returns a false positive in the scenario that at least one handler exists for that event and `undefined` was passed for the options