Use Promise.resolve().then() instead of setTimeout().
setTimeout(callback, 0) creates a macrotask, whereas Promise.resolve().then(callback) creates a microtask, and they both appear to work properly, as far as "turbo:load" is concerned.
This also alleviates potential warning messages in the console in the case that the callback of setTimeout took a significant amount of time to execute.
The setTimeout version can be used via the `options` parameter by setting `timeout` to `'setTimeout'`. The `noTimeout` option has been changed to this setting.

Signed-off-by: Brian Shenk <40068887+bashenk@users.noreply.github.com>