Reverted default `timeout` to `setTimeout`. Added `noWarn` option. Made `url` a `URL`. Bumped version.
Added `noWarn` option and removed `wkof.turbo.silenceWarnings` global configurable.
The `url` parameter passed to the listener is now a `URL` object.
The warning logged to the console is now a custom Error object, which facilitates providing the stack trace.
Bumped the version to 4.2.0.
Added many links to the README.
Updated many descriptions and spacings in the README table.
Updated the examples to match the new options and changes.

Signed-off-by: Brian Shenk <40068887+bashenk@users.noreply.github.com>